--- .env
@@ -1,41 +1,0 @@
-# Python bytecode files
-*.pyc
-*.pyo
-__pycache__/
-
-# Python virtual environment
-venv/
-env/
-.venv/
-ENV/
-
-# IDE/Editor files
-.vscode/
-.idea/
-*.swp
-*.swo
-
-# Distribution/Packaging
-*.egg
-*.egg-info/
-dist/
-build/
-*.whl
-
-# Jupyter Notebook checkpoints
-.ipynb_checkpoints/
-
-# Test coverage reports
-.coverage
-.coverage.*
-nosetests.xml
-test_*.xml
-
-# Pytest cache
-.pytest_cache/
-
-# PyCharm project files
-*.iml
-
-# Flask / Django secret key
-*.env
