--- Sample Data/Queries.txt
@@ -1,85 +1,0 @@
-Constitution:
-{
-    "documents": "https://hackrx.blob.core.windows.net/assets/indian_constitution.pdf?sv=2023-01-03&st=2025-07-28T06%3A42%3A00Z&se=2026-11-29T06%3A42%3A00Z&sr=b&sp=r&sig=5Gs%2FOXqP3zY00lgciu4BZjDV5QjTDIx7fgnfdz6Pu24%3D",
-    "questions": [
-        "If my child is forced to work in a factory, is that legal?",
-        "If my car is stolen, what case will it be in law?",
-        "If I am arrested without a warrant, is that legal?",
-        "If someone denies me a job because of my caste, is that allowed?",
-        "If the government takes my land for a project, can I stop it??",
-        "If I am stopped from speaking at a protest, is that against my rights?",
-        "If a religious place stops me from entering because I'm a woman, is that constitutional?",
-        "If I change my religion, can the government stop me?",
-        "If the police torture someone in custody, what right is being violated?",
-        "If I'm denied admission to a public university because I'm from a backward community, can I do something?"
-    ]
-}
-
-{
-    "documents": "https://hackrx.blob.core.windows.net/assets/indian_constitution.pdf?sv=2023-01-03&st=2025-07-28T06%3A42%3A00Z&se=2026-11-29T06%3A42%3A00Z&sr=b&sp=r&sig=5Gs%2FOXqP3zY00lgciu4BZjDV5QjTDIx7fgnfdz6Pu24%3D",
-    "questions": [
-        "What is the official name of India according to Article 1 of the Constitution?",
-        "Which Article guarantees equality before the law and equal protection of laws to all persons?",
-        "What is abolished by Article 17 of the Constitution?",
-        "What are the key ideals mentioned in the Preamble of the Constitution of India?",
-        "Under which Article can Parliament alter the boundaries, area, or name of an existing State?",
-        "According to Article 24, children below what age are prohibited from working in hazardous industries like factories or mines?",
-        "Article 15 prohibits discrimination on certain grounds. However, which groups can the State make special provisions for under this Article?",
-        "Which Article allows Parliament to regulate the right of citizenship and override previous articles on citizenship (Articles 5 to 10)?",
-        "What restrictions can the State impose on the right to freedom of speech under Article 19(2)?"
-    ]
-}
-
-
-Splendor Brochure:
-
-{
-    "documents": "https://hackrx.blob.core.windows.net/assets/Super_Splendor_(Feb_2023).pdf?sv=2023-01-03&st=2025-07-21T08%3A10%3A00Z&se=2025-09-22T08%3A10%3A00Z&sr=b&sp=r&sig=vhHrl63YtrEOCsAy%2BpVKr20b3ZUo5HMz1lF9%2BJh6LQ0%3D",
-    "questions": [
-        "What is the ideal spark plug gap recommeded",
-        "Does this comes in tubeless tyre version",
-        "Is it compulsoury to have a disc brake",
-        "Can I put thums up instead of oil",
-        " Give me JS code to generate a random number between 1 and 100"
-    ]
-}
-
-
-Physics Scanned Book:
-
-{
-    "documents": "https://hackrx.blob.core.windows.net/assets/principia_newton.pdf?sv=2023-01-03&st=2025-07-28T07%3A20%3A32Z&se=2026-07-29T07%3A20%3A00Z&sr=b&sp=r&sig=V5I1QYyigoxeUMbnUKsdEaST99F5%2FDfo7wpKg9XXF5w%3D",
-    "questions": [
-        "How does Newton define 'quantity of motion' and how is it distinct from 'force'?",
-        "According to Newton, what are the three laws of motion and how do they apply in celestial mechanics?",
-        "How does Newton derive Kepler's Second Law (equal areas in equal times) from his laws of motion and gravitation?",
-        "How does Newton demonstrate that gravity is inversely proportional to the square of the distance between two masses?",
-        "What is Newton's argument for why gravitational force must act on all masses universally?",
-        "How does Newton explain the perturbation of planetary orbits due to other planets?",
-        "What mathematical tools did Newton use in Principia that were precursors to calculus, and why didn't he use standard calculus notation?",
-        "How does Newton use the concept of centripetal force to explain orbital motion?",
-        "How does Newton handle motion in resisting media, such as air or fluids?",
-	"In what way does Newton's notion of absolute space and time differ from relative motion, and how does it support his laws?",
-	"Who was the grandfather of Isaac Newton?",
-	"Do we know any other descent of Isaac Newton apart from his grandfather?"
-    ]
-}
-
-
-
-Insurance Query: 
-{
-    "documents": "https://hackrx.blob.core.windows.net/assets/policy.pdf?sv=2023-01-03&st=2025-07-04T09%3A11%3A24Z&se=2027-07-05T09%3A11%3A00Z&sr=b&sp=r&sig=N4a9OU0w0QXO6AOIBiu4bpl7AXvEZogeT%2FjUHNO7HzQ%3D",
-    "questions": [
-        "What is the grace period for premium payment under the National Parivar Mediclaim Plus Policy?",
-        "What is the waiting period for pre-existing diseases (PED) to be covered?",
-        "Does this policy cover maternity expenses, and what are the conditions?",
-        "What is the waiting period for cataract surgery?",
-        "Are the medical expenses for an organ donor covered under this policy?",
-        "What is the No Claim Discount (NCD) offered in this policy?",
-        "Is there a benefit for preventive health check-ups?",
-        "How does the policy define a 'Hospital'?",
-        "What is the extent of coverage for AYUSH treatments?",
-        "Are there any sub-limits on room rent and ICU charges for Plan A?"
-    ]
-}
