+++ app/config/config.json
@@ -1,0 +1,4 @@
+{
+  "EMBEDDINGS_DIR": "embeddings",
+  "VALID_API_KEY": "2931609bd36ec1a45cb577b3b831dc711c76ae157b3c6250c564284c93b062ff"
+}
